+++ ClassLibrary1/Services/Types/Commands/DeleteTypes/DeleteTypesCommandsHandler.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace BLL.Services
 {
-    public class DeleteTypesCommandsHandler : IRequestHandler<DeleteTypesCommands>
+    public class DeleteTypesCommandsHandler : IRequestHandler<DeleteTypesCommands, APIResponse>
     {
         private readonly IUnitOfWork uow;
         private readonly IMemoryCache _cache;
@@ -17,13 +17,38 @@ namespace BLL.Services
             this.uow = uow;
             _cache = cache;
         }
-        public async Task<Unit> Handle(DeleteTypesCommands request, CancellationToken cancellationToken)
+        public async Task<APIResponse> Handle(DeleteTypesCommands request, CancellationToken cancellationToken)
         {
-            string Key = $"member-allTypesAll";
-            var DeleteItem = await uow.Types.GetByIdAsync(x => x.Id == request.Id);
-            await uow.Types.DeleteAsync(DeleteItem);
-            _cache.Remove(Key);
-            return Unit.Value;
+            try
+            {
+                string Key = $"member-allTypesAll";
+                var DeleteItem = await uow.Types.GetByIdAsync(x => x.Id == request.Id);
+                if (DeleteItem == null)
+                {
+                    return new APIResponse
+                    {
+                        IsError = false,
+                        Code = 404,
+                        Message = "Element Not Found"
+                    };
+                }
+                await uow.Types.DeleteAsync(DeleteItem);
+                _cache.Remove(Key);
+                return new APIResponse
+                {
+                    IsError = true,
+                    Code = 200,
+                    Message = "Element Deleted"
+                };
+            }
+            catch (Exception ex)
+            {
+                return new APIResponse
+                {
+                    IsError = true,
+                    Message = ex.Message,
+                };
+            }
         }
     }
 }
